+++ test/jest/__mock__/stripesCore.mock.js
@@ -3,7 +3,7 @@ jest.mock('@folio/stripes/core', () => ({
   useOkapiKy: jest.fn().mockReturnValue({
     get: jest.fn().mockReturnValue({ json: jest.fn().mockResolvedValue({}) }),
     post: jest.fn().mockReturnValue({ json: jest.fn().mockResolvedValue({}) }),
-    put: jest.fn(),
+    put: jest.fn().mockReturnValue({ json: jest.fn().mockResolvedValue({}) }),
     delete: jest.fn(),
     extend: jest.fn().mockReturnValue(this),
   }),
@@ -41,6 +41,8 @@ jest.mock('@folio/stripes/core', () => ({
   TitleManager: jest.fn(() => <div />),
   useStripes: jest.fn(() => ({
     connect: Component => Component,
+    timezone: 'UTC',
+    locale: 'en-US',
     hasInterface: jest.fn().mockReturnValue(true),
     hasPerm: jest.fn().mockReturnValue(true),
   })),